ps script to handle windows update agent on ms windows
with this powershell script it is possible to control the local agent on the computer. In principle, so updates can be installed without a race condition leads to a break out or other unpleasant side-effects. this script was created for the it-system management. it was used in conjunction with a deployment system.