migrate GPOs from Domain to Domain
It turns out that Group Policy Management Console (GPMC) has a dandy EXPORT, IMPORT feature.  The only thing that isn’t particularly obvious is that you have to EXPORT (and IMPORT) in the GROUP POLICY OBJECTS folder, and not on any of the OU’s. 
And the simpelst Way is to run Script for this Action.