Active Server Process Control PC PS Module
With this Powershell Module it can be controlled workstations remotely. The only prerequisite is administrative rights on the target machine. or the machines are in the active directory and the administrative pc as well and by group membership the permissions are given. The advantage of this module over a script is. it only has to be loaded once into a powershell instance. then it can provide all functions within the existing session. And these functions can be automated with task schedular or autorun or group policies.